% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{AvgGroupExpr}
\alias{AvgGroupExpr}
\title{Compute average (scaled) expression of cells in each group}
\usage{
AvgGroupExpr(m, groups, do.scale = F, scale.max = 6, min.cells = 3)
}
\arguments{
\item{m}{Expression matrix}

\item{groups}{Perturbation dictionary or vector}

\item{do.scale}{Scale expression matrix before averaging}

\item{scale.max}{If do.scale is True, maximum scaled value}

\item{min.cells}{Minimum cells per group to average over}
}
\value{
Matrix of average expression per group
}
\description{
Compute average (scaled) expression of cells in each group
}
