% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{CalcGeneModuleEffect}
\alias{CalcGeneModuleEffect}
\title{Calculate effect of perturbations on gene modules}
\usage{
CalcGeneModuleEffect(
  coefs.matrix,
  gene.clusters.list,
  gene.module.mapping = NULL,
  min.coef = 0.025
)
}
\arguments{
\item{coefs.matrix}{Regression coefficients matrix}

\item{gene.clusters.list}{List of gene clusters}

\item{gene.module.mapping}{Mapping gene modules to descriptions}

\item{min.coef}{Minimum cofficient magnitude for a gene to be considered}
}
\value{
Matrix of perturbation effects on each gene modules
}
\description{
Calculate effect of perturbations on gene modules
}
