% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{CalcGlmnet}
\alias{CalcGlmnet}
\title{Calculate penalized linear regression with glmnet}
\usage{
CalcGlmnet(design.matrix, metadata, y, alpha, lambda.use, lambda.seq, family,
  ctrl = NULL)
}
\arguments{
\item{design.matrix}{Design matrix for regression}

\item{metadata}{Additional covariates to take into account (will not be permuted)}

\item{y}{Linear model response}

\item{alpha}{Elasticnet ratio: (0 is fully L2, 1 fully is L1)}

\item{lambda.use}{Coefficient regularization parameter}

\item{lambda.seq}{Regularization sequence to follow}

\item{family}{Regression family to use}

\item{ctrl}{Control variable in the design matrix (optional)}
}
\value{
Matrix of regression coefficients
}
\description{
Calculate penalized linear regression with glmnet
}
