% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{CalcGlmnetPvals}
\alias{CalcGlmnetPvals}
\title{Calculate regression coefficients and p-values via permutation testing}
\usage{
CalcGlmnetPvals(design.matrix, metadata, y, alpha, lambda, family, n.rand,
  n.cores, n.bins = 10, use.quantiles = T, output.name = "cf")
}
\arguments{
\item{design.matrix}{Design matrix for regression}

\item{metadata}{Additional covariates to take into account (will not be permuted)}

\item{y}{Linear model response}

\item{alpha}{Elasticnet ratio: (0 is fully L2, 1 fully is L1)}

\item{lambda}{Coefficient regularization parameter}

\item{family}{Regression family to use}

\item{n.rand}{Number of permutations for calculating coefficient significance}

\item{n.cores}{Number of cores to use}

\item{n.bins}{If binning genes, number of bins to use}

\item{use.quantiles}{If binning genes, whether or not to bin by quantile}

\item{output.name}{Column name of regression coefficient}
}
\value{
If binning genes, returns a dataframe with Gene, Perturbation, Regression Coefficient (cf), and P-value.
        Otherwise returns a list of regression coefficients and p-values
}
\description{
Calculate regression coefficients and p-values via permutation testing
}
