% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{CalcLogFC}
\alias{CalcLogFC}
\title{Calculate log-FC for each genotype against the control}
\usage{
CalcLogFC(X, Y, ctrl, eps = 0.001)
}
\arguments{
\item{X}{Design matrix}

\item{Y}{Gene expression}

\item{ctrl}{Control genotype}
}
\value{
log-FC matrix
}
\description{
Calculate log-FC for each genotype against the control
}
