% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{CalcSNN}
\alias{CalcSNN}
\title{Build Shared Nearest Neighbors matrix. From Seurat (https://github.com/satijalab/seurat)}
\usage{
CalcSNN(data.use, k.param = 10, prune.SNN = 1/15)
}
\arguments{
\item{data.use}{Cells x features matrix to construct SNN with}

\item{k.param}{Number of neighbors}

\item{prune.SNN}{Min fraction of shared nearest neighbors to consider}

\item{k.scale}{Neighborhood size to consider}
}
\value{
Sparse shared nearest neighbors matrix
}
\description{
Build Shared Nearest Neighbors matrix. From Seurat (https://github.com/satijalab/seurat)
}
