% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{CleanDesignCtrl}
\alias{CleanDesignCtrl}
\title{Filters out cells that belong to control genotype and other genotypes
CleanDesignCtrl is deprecated, use CleanControls instead}
\usage{
CleanDesignCtrl(design.mat, ctrl)
}
\description{
Filters out cells that belong to control genotype and other genotypes
CleanDesignCtrl is deprecated, use CleanControls instead
}
