% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{ClusterWiseDE}
\alias{ClusterWiseDE}
\title{Compute per-cluster differential expression between conditions}
\usage{
ClusterWiseDE(
  obj,
  cluster.var,
  condition.var,
  condition.1,
  condition.2,
  min.cells = 10,
  ...
)
}
\arguments{
\item{obj}{Seurat object}

\item{cluster.var}{Name of cluster metadata column}

\item{condition.var}{Name of condition metadata column}

\item{condition.1}{Name of first condition to compare}

\item{condition.2}{Name of second condition to compare}

\item{min.cells}{Minimum cells per cluster/condition combination}

\item{...}{Additional keyword parameters to pass to FindMarkers}
}
\value{
Dataframe with per-cluster DE results
}
\description{
Compute per-cluster differential expression between conditions
}
