% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phemd.R
\name{computeEMD}
\alias{computeEMD}
\title{Visualize correlation between two vectors.
Adapted from PhEMD (https://bioconductor.org/packages/release/bioc/html/phemd.html)}
\usage{
computeEMD(pheno.dict, clusters, clusters.dists)
}
\arguments{
\item{pheno.dict}{Named list of cell vectors for each genotype}

\item{clusters}{Cell type assignments}

\item{clusters.dists}{Matrix of distances between cell type centroids}
}
\value{
Earth mover's distance between all genotypes
}
\description{
Visualize correlation between two vectors.
Adapted from PhEMD (https://bioconductor.org/packages/release/bioc/html/phemd.html)
}
