% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{CreateDesignMatrix}
\alias{CreateDesignMatrix}
\title{Convert perturbation dictionary to design matrix}
\usage{
CreateDesignMatrix(groups, min.cells = 10)
}
\arguments{
\item{groups}{Perturbation dictionary (in list format) or named vector}

\item{min.cells}{Min number of cells per perturbation}
}
\value{
Sparse binary matrix where rows are cells, and columns are perturbation. 1 means the cell received the perturbation.
}
\description{
Convert perturbation dictionary to design matrix
}
