% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{cross_validate_lambda}
\alias{cross_validate_lambda}
\title{Helper function for running cross validation on a sequence of lambda values}
\usage{
cross_validate_lambda(x, y, groups.list, alpha, metric, family, lambda, folds)
}
\arguments{
\item{x}{Design matrix + covariates matrix}

\item{y}{Expression response}

\item{groups.list}{Perturbation dictionary}

\item{alpha}{Alpha value}

\item{metric}{Metric to use for evaluation (pearson, spearman)}

\item{family}{GLM family to use for elasticnet (default: mgaussian)}

\item{lambda}{lambda values to test}

\item{folds}{List of train/test splits}
}
\value{
Matrix of cross validation errors per fold
}
\description{
Helper function for running cross validation on a sequence of lambda values
}
