% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{EvaluateLinearModel}
\alias{EvaluateLinearModel}
\title{Function for evaluating linear model performance}
\usage{
EvaluateLinearModel(
  mdl.fit,
  x.test,
  y.test,
  groups,
  eval.metric,
  do.scale = T,
  min.cells = 3
)
}
\arguments{
\item{mdl.fit}{Linear model fit on training data}

\item{x.test}{Test design matrix}

\item{y.test}{Test response}

\item{groups}{Perturbation dictionary (in list format) or named vector}

\item{eval.metric}{Model evaluation metric}

\item{do.scale}{Whether to scale response before evaluation (only applies to pearson & spearman metrics)}

\item{min.cells}{Minimum cells per group to include in evaluation}
}
\description{
Function for evaluating linear model performance
}
