% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{ExtractField}
\alias{ExtractField}
\title{Extract a field from a delimited string}
\usage{
ExtractField(string, field = 1, delim = "_")
}
\arguments{
\item{string}{Input string}

\item{field}{Field to extract}

\item{delim}{Character delimiter (default: "_")}
}
\description{
Extract a field from a delimited string
}
