% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation.R
\name{FastRead}
\alias{FastRead}
\title{Efficiently load tabular datasets using data.table}
\usage{
FastRead(fi, sep = "\\t")
}
\arguments{
\item{fi}{File path}

\item{sep}{Delimiter character (default tab)}
}
\description{
Efficiently load tabular datasets using data.table
}
