% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FilterData}
\alias{FilterData}
\title{Filter rows and columns of counts matrix}
\usage{
FilterData(counts, min.cells.frac, trim, min.genes = 500, min.expr = 0,
  max.transcripts = 50000)
}
\arguments{
\item{counts}{Input counts matrix}

\item{min.cells.frac}{Min number of cells to keep a gene}

\item{min.genes}{Min genes expressed to keep a cell}

\item{min.expr}{Minimum level to count as expressed}

\item{max.transcripts}{Max number of UMIs per cell}
}
\value{
Filtered counts matrix
}
\description{
Filter rows and columns of counts matrix
}
