% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{FisherEnrich}
\alias{FisherEnrich}
\title{Use Fisher's exact test to calculate enrichment for count data}
\usage{
FisherEnrich(marker.genes, genesets, n.background, correct = F)
}
\arguments{
\item{marker.genes}{Top marker genes}

\item{genesets}{List of genesets}

\item{n.background}{Number of background genes}

\item{correct}{FDR correction}
}
\value{
Dataframe with genesets, Fisher enrichment p-values (p.val), number of genes in the geneset (genes_in_set),
        size of each geneset (geneset_size)
}
\description{
Use Fisher's exact test to calculate enrichment for count data
}
