% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_handling.R
\name{FlattenGenotypeList}
\alias{FlattenGenotypeList}
\title{Convert genotypes dictionary from list to named vector. All cells with more than one genotype are removed}
\usage{
FlattenGenotypeList(genotypes.list)
}
\arguments{
\item{genotypes.list}{Named list of cell vectors for each genotype}
}
\value{
Named vector with the values as genotypes and the names as cell names.
}
\description{
Convert genotypes dictionary from list to named vector. All cells with more than one genotype are removed
}
