% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GenesetDotPlot}
\alias{GenesetDotPlot}
\title{Dot plot visualization}
\usage{
GenesetDotPlot(data.plot.df, color.col, size.col, row.col = "Geneset",
  col.col = "Group", cols.use = c("lightgrey", "blue"), col.max = 4,
  dot.scale = 4, scale.by = "radius", scale.min = NA,
  scale.max = NA, plot.legend = F, font.size = 12)
}
\arguments{
\item{data.plot.df}{Dataframe with Geneset, Group, and enrichment info}

\item{color.col}{data.plot.df column to determine dot color}

\item{size.col}{data.plot.df column to determine dot size}

\item{row.col}{Plot rows (usually genesets)}

\item{col.col}{Plot columns (usually groups)}

\item{cols.use}{Colors to plot, can pass a single character giving the name of
a palette from \code{RColorBrewer::brewer.pal.info}}

\item{col.max}{Maximum color value to set}

\item{dot.scale}{Scale the size of the points, similar to cex}

\item{scale.by}{Scale the size of the points by 'size' or by 'radius'}

\item{scale.min}{Set lower limit for scaling, use NA for default}

\item{scale.max}{Set upper limit for scaling, use NA for default}

\item{plot.legend}{plots the legends}

\item{font.size}{x and y axis font sizes}

\item{dot.min}{The fraction of cells at which to draw the smallest dot
(default is 0). All cell groups with less than this expressing the given
gene will have no dot drawn.}
}
\value{
returns a ggplot2 object
}
\description{
Intuitive way of visualizing geneset enrichment across different
groups (genotypes, clusters). The size of the dot encodes the percentage of
gsea enrichment within a group, while the color encodes the Fisher enrichment
cells within a group (blue is high).
}
