% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{GenotypeClusterChisq}
\alias{GenotypeClusterChisq}
\title{Calculate chi-squared test of difference between each row of a counts matrix and the specified control row}
\usage{
GenotypeClusterChisq(counts.df, ntc)
}
\arguments{
\item{counts.df}{Counts matrix}

\item{ntc}{Control row}
}
\value{
Vector of p-values for each row
}
\description{
Calculate chi-squared test of difference between each row of a counts matrix and the specified control row
}
