% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_handling.R
\name{GenotypeClusterCounts}
\alias{GenotypeClusterCounts}
\title{Counts the number of cells in each genotype x cluster combination}
\usage{
GenotypeClusterCounts(genotypes.list, clusters.list)
}
\arguments{
\item{genotypes.list}{Named list of cell vectors for each genotype}

\item{clusters.list}{Named list of cell vectors for each cluster}
}
\value{
Matrix of genotypes and cluster counts
}
\description{
Counts the number of cells in each genotype x cluster combination
}
