% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggheat}
\alias{ggheat}
\title{Plot matrix as heatmap. Options for clustering rows/columns}
\usage{
ggheat(m, rescaling = "none", clustering = "none", labCol = T,
  labRow = T, border = F, heatscale = c(low = "skyblue", mid = "white",
  high = "tomato"), legend.title = NULL, x.lab.size = 8, y.lab.size = 8)
}
\arguments{
\item{m}{Input matrix to visualize}

\item{rescaling}{Rescale the matrix. Options include row/column wise scaling, or a custom scaling function}

\item{clustering}{Cluster rows/columns or both}

\item{labCol}{Label columns}

\item{labRow}{Label rows}

\item{border}{Plot heatmap border}

\item{heatscale}{Colormap to use}

\item{legend.title}{Colormap title}

\item{x.lab.size}{X-axis font size}

\item{y.labl.size}{Y-axis font size}
}
\value{
ggplot2 object with heatmap plot
}
\description{
Plot matrix as heatmap. Options for clustering rows/columns
}
