% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{GroupClusterCounts}
\alias{GroupClusterCounts}
\title{Counts the number of cells in each perturbation/cluster}
\usage{
GroupClusterCounts(groups, clusters)
}
\arguments{
\item{groups}{Perturbation dictionary (in list format) or named vector}

\item{clusters}{Cluster annotations}
}
\value{
Dataframe of cluster cell counts for each perturbation
}
\description{
Counts the number of cells in each perturbation/cluster
}
