% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{GroupOverlapCounts}
\alias{GroupOverlapCounts}
\title{Counts the number of cells in each group overlap combination}
\usage{
GroupOverlapCounts(group.list.1, group.list.2)
}
\arguments{
\item{group.list.1}{Named list of cell vectors for each genotype}

\item{group.list.2}{Named list of cell vectors for each cluster}
}
\value{
Matrix of cell counts that overlap the two sets of groups
}
\description{
Counts the number of cells in each group overlap combination
}
