% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{LoadGenesets}
\alias{LoadGenesets}
\title{Read genesets from gmt file to list format}
\usage{
LoadGenesets(file.name)
}
\arguments{
\item{file.name}{Geneset file name}
}
\value{
List of genesets
}
\description{
Read genesets from gmt file to list format
}
