% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{MergeGroupLists}
\alias{MergeGroupLists}
\title{Merge two perturbation dicts}
\usage{
MergeGroupLists(group.list.1, group.list.2)
}
\arguments{
\item{group.list.1}{1st Perturbation dictionary to merge}

\item{group.list.2}{1st Perturbation dictionary to merge}
}
\value{
Merged perturbation dictionary
}
\description{
Merge two perturbation dicts
}
