% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{MergeGroups}
\alias{MergeGroups}
\title{Merge perturbations with the same pattern in names.
For example, we might want to collapse all guide RNAs into the gene they
target (i.e. TULP2-1, TULP2-2, TULP2-3 into TULP2)}
\usage{
MergeGroups(group.list, pattern, new.group.name = NULL)
}
\arguments{
\item{group.list}{Perturbation dictionary (in list format)}

\item{pattern}{Character pattern to collapse (i.e. TULP2)}

\item{new.group.name}{Name of new perturbation (default: pattern)}
}
\value{
List of perts with all perturbations containing the pattern merged
}
\description{
Merge perturbations with the same pattern in names.
For example, we might want to collapse all guide RNAs into the gene they
target (i.e. TULP2-1, TULP2-2, TULP2-3 into TULP2)
}
