% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{MouseHumanMapping}
\alias{MouseHumanMapping}
\title{Map mouse genes to human genes}
\usage{
MouseHumanMapping(mouse.genes, mouse.human.gene.file)
}
\arguments{
\item{mouse.genes}{Vector of mouse genes}

\item{mouse.human.gene.file}{File with mouse to human gene mappings}
}
\value{
vector of human genes with mouse genes as names
}
\description{
Map mouse genes to human genes
}
