% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{MultipleFisherEnrich}
\alias{MultipleFisherEnrich}
\title{Count over-enrichment for a list of gene markers}
\usage{
MultipleFisherEnrich(markers.list, genesets, correct = F)
}
\arguments{
\item{markers.list}{List of top marker genes}

\item{genesets}{List of genesets}

\item{correct}{FDR correction}
}
\value{
Dataframe with group, genesets, Fisher enrichment p-values (p.val),
        number of genes in the geneset (genes_in_set),
        size of each geneset (geneset_size)
}
\description{
Count over-enrichment for a list of gene markers
}
