% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{MultipleGSEAEnrich}
\alias{MultipleGSEAEnrich}
\title{GSEA over-enrichment test using liger}
\usage{
MultipleGSEAEnrich(
  scores.list,
  genesets,
  n.rand = 1000,
  n.cores = 1,
  power = 1,
  quantile.threshold = min(100/n.rand, 0.1)
)
}
\arguments{
\item{scores.list}{List of scores to test enrichment for}

\item{genesets}{List of genesets}

\item{n.rand}{Number of randomizations}

\item{n.cores}{Number of cores}

\item{power}{GSEA exponent}

\item{quantile.threshold}{Quantile threshold}
}
\value{
Dataframe with GSEA enrichment
}
\description{
GSEA over-enrichment test using liger
}
