% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_handling.R
\name{PadDesignMatrix}
\alias{PadDesignMatrix}
\title{Pad design matrix with zeros for cells without a called genotype}
\usage{
PadDesignMatrix(X, row.names)
}
\arguments{
\item{X}{Design matrix}

\item{row.names}{Rownames (must contain at least all rows of X)}
}
\value{
Design matrix, with all row.names not in X set to all zeros
}
\description{
Pad design matrix with zeros for cells without a called genotype
}
