% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotBox}
\alias{PlotBox}
\title{Boxplot of features}
\usage{
PlotBox(
  obj,
  feature,
  group.by,
  fill.by,
  group.subset = NULL,
  fill.subset = NULL,
  pt.size = 0.75
)
}
\arguments{
\item{obj}{Seurat object}

\item{feature}{Feature to plot}

\item{group.by}{Name of main grouping factor (the x-axis of boxplot)}

\item{fill.by}{Name of secondary grouping factor (fill color of the boxplot)}

\item{group.subset}{Plot only a subset of groups (default: NULL)}

\item{fill.subset}{Plot only a subset of fill groups (default: NULL)}

\item{pt.size}{Point size (set to zero to hide points)}
}
\value{
Boxplot of expression values
}
\description{
Boxplot of features
}
