% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotClusterComp}
\alias{PlotClusterComp}
\title{Plot cluster compositions}
\usage{
PlotClusterComp(
  group.list,
  clusters,
  samples = NULL,
  stabilize.var = TRUE,
  ctrl.group = NULL,
  groups.plot = NULL,
  clusters.plot = NULL,
  pt.size = 0.75
)
}
\arguments{
\item{group.list}{Perturbation dictionary (in list format)}

\item{clusters}{Cluster annotations}

\item{samples}{Samples (i.e. mice, patients, biological replicates, default: NULL)}

\item{stabilize.var}{Apply variance stabilizing transformation (default: TRUE)}

\item{ctrl.group}{Control perturbation to compute logFC against (for paired samples only, default: NULL)}

\item{groups.plot}{Subset of perturbations to plot (default: NULL)}

\item{clusters.plot}{Subset of clusters to plot (default: NULL)}

\item{pt.size}{Point size (set to zero to hide points)}
}
\value{
Boxplot of cluster compositions
}
\description{
Plot cluster compositions
}
