% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotDims}
\alias{PlotDims}
\title{Visualize t-SNE or PCA plots}
\usage{
PlotDims(dim.x, dim.y, clusters = NULL, x.lab = "tsne1", y.lab = "tsne2",
  main.title = NULL, pt.size = 1, font.size = 12, alpha = 1,
  do.label = T, label.size = 4, show.legend = T, label.points = F)
}
\arguments{
\item{dim.x}{X-axis vector}

\item{dim.y}{Y-axis vector}

\item{clusters}{Cell clusters}

\item{x.lab}{X-axis title}

\item{y.lab}{Y-axis title}

\item{main.title}{Plot title}

\item{pt.size}{Data point size}

\item{font.size}{Font size for axis titles}

\item{alpha}{Data point transparency}

\item{do.label}{Whether or not to label the clusters on the plot}

\item{label.size}{Size of cluster labels}

\item{show.legend}{Display legend}

\item{label.points}{Label each individual points}
}
\value{
ggplot2 object
}
\description{
Visualize t-SNE or PCA plots
}
