% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotHexBin}
\alias{PlotHexBin}
\title{Visualize correlation between two vectors with density plot.
Useful for visualizing large datasets.}
\usage{
PlotHexBin(v, w, n.bins = 100)
}
\arguments{
\item{v}{X-axis vector}

\item{w}{Y-axis vector}

\item{n.bins}{Number of hexagonal bins}
}
\value{
ggplot2 object with hexbin plot
}
\description{
Visualize correlation between two vectors with density plot.
Useful for visualizing large datasets.
}
