% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotResiduals}
\alias{PlotResiduals}
\title{Plot predicted values vs residuals to assess data linearity and heteroskedasticity
This requires densifying the expression matrices so it might be wise to only plot a subset of cells/genes.}
\usage{
PlotResiduals(x, y, mfit, nonzero.only = T)
}
\arguments{
\item{x}{Design matrix + covariates matrix}

\item{y}{Expression response}

\item{mfit}{Fitted glmnet model}

\item{nonzero.only}{Only plot residuals vs predicted for nonzero values}
}
\value{
KDE density plot of residuals vs predicted
}
\description{
Plot predicted values vs residuals to assess data linearity and heteroskedasticity
This requires densifying the expression matrices so it might be wise to only plot a subset of cells/genes.
}
