% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotViolin}
\alias{PlotViolin}
\title{Violin plot (adapted from Seurat)}
\usage{
PlotViolin(feature, data, cell.ident, do.sort = F, y.max = NULL,
  size.x.use = 10, size.y.use = 10, x.title = "Group",
  size.title.use = 11, point.size.use = 1, cols.use = NULL, y.log = F,
  x.lab.rot = F, y.lab.rot = F, legend.position = "right",
  remove.legend = F)
}
\arguments{
\item{feature}{Feature name}

\item{data}{Numeric vector of data to plot}

\item{cell.ident}{Factor with cell identities (must be same length as data)}

\item{do.sort}{Sort factor levels}

\item{y.max}{Set max expression level}

\item{size.x.use}{X axis text font size}

\item{size.y.use}{Y axis text font size}

\item{x.title}{X-axis title}

\item{size.title.use}{X & Y axis title font sizes}

\item{point.size.use}{Jitter point size}

\item{cols.use}{Color palette to use}

\item{y.log}{Log-scale gene expression}

\item{x.lab.rot}{Rotate x-axis labels}

\item{y.lab.rot}{Rotate y-axis labels}

\item{legend.position}{Legend position ("right" or "left")}

\item{remove.legend}{If true, hides legend}
}
\value{
A violin plot
}
\description{
Violin plot (adapted from Seurat)
}
