% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{PredictMixedLM}
\alias{PredictMixedLM}
\title{Predict expression given a new model dataframe}
\usage{
PredictMixedLM(fit_list, new_model_df)
}
\arguments{
\item{fit_list}{List of fitted glmmLasso objects}

\item{new_model_df}{New model dataframe to predict on (must contain same covariate columns as original dataframe)}
}
\value{
Matrix of predicted values
}
\description{
Predict expression given a new model dataframe
}
