% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ReadData}
\alias{ReadData}
\title{Read data from either a tab separated values file or a 10X genomics sparse matrix directory}
\usage{
ReadData(matrix.dir)
}
\arguments{
\item{matrix.dir}{Input matrix file or directory}
}
\value{
Matrix
}
\description{
Read data from either a tab separated values file or a 10X genomics sparse matrix directory
}
