% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_handling.R
\name{ReadGenotypes}
\alias{ReadGenotypes}
\title{Read in genotypes dictionary from csv file. Returns an R list.}
\usage{
ReadGenotypes(pheno.dict.file, sep.char = ",")
}
\arguments{
\item{pheno.dict.file}{Genotype dictionary file in csv format}

\item{sep.char}{Value separating columns in genotype dictionary file}
}
\value{
Named list of cell vectors for each genotype
}
\description{
Read in genotypes dictionary from csv file. Returns an R list.
}
