% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{RunCrossValidation}
\alias{RunCrossValidation}
\title{Run cross validation for both alpha and lambda}
\usage{
RunCrossValidation(
  x,
  y,
  groups,
  metric = "pearson",
  alpha.seq = c(0.1, 0.4, 0.7, 0.9, 0.95, 1),
  plot = T,
  n.cores = 4,
  family = "mgaussian",
  nlambda = 10,
  lambda.min.ratio = 0.01,
  nfolds = 4,
  seed = NULL,
  seq.lambda.pred = F
)
}
\arguments{
\item{x}{Design matrix + covariates matrix}

\item{y}{Expression response}

\item{groups}{Perturbation dictionary (in list format) or named vector}

\item{metric}{Metric to evaluate models (must be either pearson, spearman, or euclidean)}

\item{alpha.seq}{Sequence of alpha values to test}

\item{plot}{Plot cross validation results (default: True)}

\item{n.cores}{Number of cores to use (default: 4)}

\item{family}{GLM family to use for elasticnet (default: mgaussian)}

\item{nlambda}{Number of lambda values to test (default: 10)}

\item{lambda.min.ratio}{Sets the minimum lambda value to test (default: 0.01)}

\item{nfolds}{Number of folds to cross validate over (default: 5)}

\item{seed}{Random seed for fold reproducibility}

\item{seq.lambda.pred}{Predict expression at each lambda value sequentially to save memory (default: F)}
}
\value{
List of results containing cross validation objects (cv.list),
        cross validation summary stats (cv.summary),
        and the optimal lambda/alpha combo (alpha, lambda)
}
\description{
Run cross validation for both alpha and lambda
}
