% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{RunLinearModel}
\alias{RunLinearModel}
\title{Wrapper function for running a ElasticNet model on a single cell dataset}
\usage{
RunLinearModel(
  obj,
  pert.col,
  batch.col = NULL,
  size.factor.col = NULL,
  mito.col = NULL,
  response.col = NULL,
  features.use = NULL,
  test.prop = 0.2,
  family = NULL,
  alpha.seq = c(0.1, 0.4, 0.7, 0.9, 0.95, 1),
  plot.cv = T,
  n.cores = 4,
  nlambda = 10,
  lambda.min.ratio = 0.01,
  nfolds = 4,
  seed = NULL,
  seq.lambda.pred = F,
  eval.metric = NULL
)
}
\arguments{
\item{obj}{Seurat object}

\item{pert.col}{Metadata column containing perturbation info (required)}

\item{batch.col}{Metadata column containing batch info (default: NULL)}

\item{size.factor.col}{Metadata column containing library size info (default: NULL)}

\item{mito.col}{Metadata column containing mitochondrial fraction info (default: NULL)}

\item{response.col}{Metadata column containing the desired response.
If NULL, the normalized data in the Seurat object will be used as the response.}

\item{features.use}{If response.col is NULL, a subset of features to use for the response}

\item{test.prop}{Proportion of cells held out for model evaluation}

\item{family}{GLM family to use for elasticnet (default: mgaussian)}

\item{alpha.seq}{Sequence of alpha values to test}

\item{plot.cv}{Plot cross validation results (default: True)}

\item{n.cores}{Number of cores to use (default: 4)}

\item{nlambda}{Number of lambda values to test (default: 10)}

\item{lambda.min.ratio}{Sets the minimum lambda value to test (default: 0.01)}

\item{nfolds}{Number of folds to cross validate over (default: 5)}

\item{seed}{Random seed for fold reproducibility}

\item{seq.lambda.pred}{Predict expression at each lambda value sequentially to save memory (default: F)}

\item{eval.metric}{Model evaluation metric}
}
\value{
Returns a list with the design matrix, response, train/test split
        cross validation results, fitted model, model coefficients, and evaluation metrics
}
\description{
Wrapper function for running a ElasticNet model on a single cell dataset
}
