% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{RunMixedLM}
\alias{RunMixedLM}
\title{Run per-gene linear mixed models using glmmLasso}
\usage{
RunMixedLM(
  model_df,
  Y,
  fixed_form,
  rand_form,
  lambda = 0,
  fam = gaussian(),
  n_cores = 8,
  verbose = F
)
}
\arguments{
\item{model_df}{Dataframe containing covariates used in the model formulas}

\item{Y}{Gene expression matrix}

\item{fixed_form}{Model formula for fixed effects}

\item{rand_form}{Model formula for random effects}

\item{lambda}{Lasso regularization parameter (default: 0)}

\item{fam}{Distribution that describes gene expression (default: gaussian())}

\item{n_cores}{Number of cores to use for parallel processing (default: 8)}

\item{verbose}{Print error statements (default: False)}
}
\value{
List of fitted linear mixed model objects (one per gene)
}
\description{
Run per-gene linear mixed models using glmmLasso
}
