% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{SetupLinearModel}
\alias{SetupLinearModel}
\title{Set up design matrices and response given a Seurat object}
\usage{
SetupLinearModel(
  obj,
  pert.col,
  batch.col = NULL,
  size.factor.col = NULL,
  mito.col = NULL,
  response.col = NULL,
  features.use = NULL,
  test.prop = 0.2,
  seed = 12345
)
}
\arguments{
\item{obj}{Seurat object}

\item{pert.col}{Metadata column containing perturbation info (required)}

\item{batch.col}{Metadata column containing batch info (default: NULL)}

\item{size.factor.col}{Metadata column containing library size info (default: NULL)}

\item{mito.col}{Metadata column containing mitochondrial fraction info (default: NULL)}

\item{response.col}{Metadata column containing the desired response.
If NULL, the normalized data in the Seurat object will be used as the response.}

\item{features.use}{If response.col is NULL, a subset of features to use for the response}

\item{test.prop}{Proportion of cells held out for model evaluation}
}
\value{
List of x (perturbation + covariate design matrix)
        y (response vector or matrix), and family (response family)
}
\description{
Set up design matrices and response given a Seurat object
}
