% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{SplitFoldsByGroup}
\alias{SplitFoldsByGroup}
\title{Splits data into folds stratified by a perturbation dictionary}
\usage{
SplitFoldsByGroup(groups, nfolds, seed = NULL)
}
\arguments{
\item{groups}{Perturbation vector or dictionary}

\item{nfolds}{Number of folds}

\item{seed}{Set a random seed for reproducibility}
}
\value{
List of fold splits
}
\description{
Splits data into folds stratified by a perturbation dictionary
}
