% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{StratifyGroupList}
\alias{StratifyGroupList}
\title{Stratify groups by additional factor}
\usage{
StratifyGroupList(group.list, stratify, sep = "_")
}
\arguments{
\item{group.list}{Perturbation dictionary (in list format)}

\item{stratify}{Factor to stratify by}

\item{sep}{Character delimiter to add to the stratified perturbation names (default: "_")}
}
\value{
Perturbation dictionary stratified by factor
}
\description{
Stratify groups by additional factor
}
