% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output.R
\name{Write10X}
\alias{Write10X}
\title{Write sparse matrix using the 10X file format}
\usage{
Write10X(
  m,
  out.dir,
  feature_id = NULL,
  feature_short_name = NULL,
  feature_type = "Gene Expression"
)
}
\arguments{
\item{m}{Sparse matrix (features x barcodes)}

\item{out.dir}{Output directory}

\item{feature_id}{Vector of feature IDs (i.e. ENSEMBL IDs)}

\item{feature_short_name}{Vector of feature short names (i.e. gene symbols)}
}
\description{
Write sparse matrix using the 10X file format
}
