% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{WriteGenesets}
\alias{WriteGenesets}
\title{Write genesets from list to gmt file}
\usage{
WriteGenesets(genesets, file.name)
}
\arguments{
\item{genesets}{List of genesets}

\item{file.name}{GMT file name}
}
\value{
Writes genesets to file.name
}
\description{
Write genesets from list to gmt file
}
