% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_handling.R
\name{WriteGenotypes}
\alias{WriteGenotypes}
\title{Write genotypes to csv file}
\usage{
WriteGenotypes(genotypes.list, out.file)
}
\arguments{
\item{genotypes.list}{Named list of cell vectors for each genotype}

\item{out.file}{Output file name}
}
\value{
Writes genotypes.list to out.file
}
\description{
Write genotypes to csv file
}
