% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{CalcClusterEnrich}
\alias{CalcClusterEnrich}
\title{Calculate cluster enrichment with logistic regression}
\usage{
CalcClusterEnrich(design.mat, clusters, alpha, lambda.use)
}
\arguments{
\item{design.mat}{Design matrix for regression}

\item{clusters}{Cluster assignments}

\item{alpha}{Elasticnet ratio: (0 is fully L2, 1 fully is L1)}

\item{lambda.use}{Shrinkage parameter}
}
\value{
Matrix of regression coefficients
}
\description{
Calculate cluster enrichment with logistic regression
}
