% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_deprecated.R
\name{CalcGlmnetPvals}
\alias{CalcGlmnetPvals}
\title{Calculate regression coefficients and p-values via permutation testing}
\usage{
CalcGlmnetPvals(
  design.matrix,
  metadata,
  y,
  alpha,
  lambda.use,
  family,
  ctrl = NULL,
  n.rand = 20,
  n.cores = 16,
  n.bins = 10,
  use.quantiles = T,
  output.name = "cf"
)
}
\arguments{
\item{design.matrix}{Design matrix for regression}

\item{metadata}{Additional covariates to take into account (will not be permuted)}

\item{y}{Linear model response}

\item{alpha}{Elasticnet ratio: (0 is fully L2, 1 fully is L1)}

\item{lambda.use}{Coefficient regularization parameter}

\item{family}{Regression family to use}

\item{ctrl}{Control variable to compare against (optional)}

\item{n.rand}{Number of permutations for calculating coefficient significance}

\item{n.cores}{Number of cores to use}

\item{n.bins}{If binning genes, number of bins to use}

\item{use.quantiles}{If binning genes, whether or not to bin by quantile}

\item{output.name}{Column name of regression coefficient}
}
\value{
Returns a dataframe with Gene, Perturbation, log-FC, p-value
}
\description{
Calculate regression coefficients and p-values via permutation testing
}
