% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_handling.R
\name{CleanDesignCtrl}
\alias{CleanDesignCtrl}
\title{Filters out cells that belong to control genotype and other genotypes}
\usage{
CleanDesignCtrl(design.mat, ctrl)
}
\arguments{
\item{design.mat}{Design matrix}
}
\value{
Filtered design matrix
}
\description{
Filters out cells that belong to control genotype and other genotypes
}
