% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{CreateDesignMatrix}
\alias{CreateDesignMatrix}
\title{Convert perturbation dictionary to design matrix}
\usage{
CreateDesignMatrix(group.list, max.groups = 3, min.cells = 10)
}
\arguments{
\item{group.list}{Named list of cell vectors for each perturbation}

\item{max.groups}{Maximum number of perturbations per cell}

\item{min.cells}{Minumum number of cells per perturbation, including combo perturbations}
}
\value{
Sparse binary matrix where rows are cells, and columns are perturbation. 1 means the cell received the perturbation.
}
\description{
Convert perturbation dictionary to design matrix
}
