% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{DesignMatrixGenotypes}
\alias{DesignMatrixGenotypes}
\title{Convert genotypes list to design matrix.
DesignMatrixGenotypes is deprecated, use CreateDesignMatrix.}
\usage{
DesignMatrixGenotypes(genotypes.list, max.genotypes = 2, min.cells = 10)
}
\description{
Convert genotypes list to design matrix.
DesignMatrixGenotypes is deprecated, use CreateDesignMatrix.
}
