% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{GetCoefMatrix}
\alias{GetCoefMatrix}
\title{Extract coefficient matrix from a multigaussian glmnet object}
\usage{
GetCoefMatrix(mfit, lambda)
}
\arguments{
\item{mfit}{Glmnet object (multigaussian)}

\item{best.lambda}{Optimal lambda regularization parameter}
}
\value{
Matrix of regression coefficients
}
\description{
Extract coefficient matrix from a multigaussian glmnet object
}
