% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{GroupClusterComp}
\alias{GroupClusterComp}
\title{Get cluster composition of each perturbation}
\usage{
GroupClusterComp(
  groups,
  clusters,
  samples = NULL,
  stabilize.var = TRUE,
  ctrl.group = NULL
)
}
\arguments{
\item{groups}{Perturbation dictionary (in list format) or vector}

\item{clusters}{Cluster annotations}

\item{samples}{Samples (i.e. mice, patients, biological replicates, default: NULL)}

\item{stabilize.var}{Apply variance stabilizing transformation (default: TRUE)}

\item{ctrl.group}{Control perturbation to compute logFC against (for paired samples only, default: NULL)}
}
\value{
Dataframe of cluster compositions or logFCs for each perturbation
}
\description{
Get cluster composition of each perturbation
}
