% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotCorrelation}
\alias{PlotCorrelation}
\title{Visualize correlation between two vectors}
\usage{
PlotCorrelation(
  x,
  y,
  x.lab = NULL,
  y.lab = NULL,
  title = NULL,
  pts.use = NULL,
  labels = NULL,
  font.size = 14,
  label.font.size = 4,
  pt.size = 1,
  x.lim = NULL,
  y.lim = NULL,
  show.corr = F,
  alpha = 1,
  pt.color = NULL
)
}
\arguments{
\item{x}{X-axis vector}

\item{y}{Y-axis vector}

\item{x.lab}{X-axis title}

\item{y.lab}{Y-axis title}

\item{title}{Plot title}

\item{pts.use}{Data points to plot}

\item{labels}{Data points to label}

\item{font.size}{Axis and plot title font size}

\item{label.font.size}{Size to use for data point labels}

\item{pt.size}{Data point size}

\item{show.corr}{Display pearson correlation in title}

\item{alpha}{Data point transparency}

\item{pt.color}{Colors to use for data points}

\item{box}{Force x and y vectors to be on the same scale}
}
\value{
ggplot2 object with correlation plot
}
\description{
Visualize correlation between two vectors
}
